third_party/WebKit/LayoutTests/external/wpt/fetch/api/response/response-trailer-expected.txt

This is a testharness.js-based test.
FAIL trailer() test assert_equals: expected (string) "X-Test-Me" but got (object) null
Harness: the test ran to completion.
